moved KingdomMemberRank out of ranks package
